app/requirements.txt
dotenv~=0.9.9
python-dotenv~=1.1.0
docker==7.1.0
kafka-python==2.2.3
